Move beef code to transaction folder, delete helpers
