--- src/beef/index.ts
@@ -1,4 +1,0 @@
-export { default as BeefTx } from './BeefTx.js'
-export { default as BeefParty } from './BeefParty.js'
-export { default as Beef } from './Beef.js'
-export { BEEF_MAGIC, BEEF_MAGIC_V2, BEEF_MAGIC_TXID_ONLY_EXTENSION } from './Beef.js'
