+++ src/transaction/index.ts
@@ -5,3 +5,4 @@ export type { default as TransactionOutput } from './TransactionOutput.js'
 export type { Broadcaster, BroadcastFailure, BroadcastResponse } from './Broadcaster.js'
 export { isBroadcastResponse, isBroadcastFailure } from './Broadcaster.js'
 export type { default as ChainTracker } from './ChainTracker.js'
+export { default as BeefTx } from './BeefTx.js'
